requirements.txt
mlx>=0.22.0
mlx-lm>=0.21.0
mlx-vlm>=0.3.12
anthropic>=0.42.0
python-dotenv>=1.0.0
pyyaml>=6.0
langgraph>=0.2.0
langchain-core>=0.3.0
openpyxl>=3.1.0
